Fix lexical errors

+++ core/src/main/java/com/trashboxbobylev/tormentpixeldungeon/actors/mobs/AdvGnoll.java
@@ -42,7 +42,7 @@ public class AdvGnoll extends Mob {
 	}
 	
     @Override
-    protected int attackDelay() {
+    protected float attackDelay() {
         return 0.33f;
     }
